i3: Config + xrandr config for work
The i3 config is almost the default one, generated from i3-config-wizard
but with the hjkl instead of jklö (on a Swedish keyboard).

The xrandr call was rendered by installing arandr, organizing the
screens and placing the output somewhere where .xprofile could read it.

All of these files should be symlinked to ~

Signed-off-by: Carl Helmertz <chelmertz@op5.com>