focus_emacs: Focus emacs by pid, not part of WM_CLASS
When firefox is viewing a page with the title "bla bla Emacs", that
window was activated before this patch. Now, we cross reference the pid
with the window listing (wmctrl -l -p <- list with pid) and use the
window's id (first column of wmctrl -l -p) to give focus to the actual
emacs window.

It feels like my shell skills may come around and bite me..