Add a timer ($mod+t)
To allow for cancelling, I needed to add `--block` and
`--timeout`, instead of sleeping in the main loop, otherwise I could not
trap the exit "signal" of dunstify.

I want to allow cancelling by ctrl+space or just left-clicking the
notification window, but after having closed that notification, the loop
continued.

This would work alright, but there's a flicker for each update, which is
super annoying.