emacs/doom: Work around clashing leader key
When I tried to open a file, the first try wouldn't "take". The error
message went something like
"Key sequence SPC m h starts with non-prefix key SPC", where SPC is the
default leader key (and, seemingly, the default ).

I'm not sure this solution is the best one, but it worked out fine so
far.